% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfmt.R
\name{install_rfmt_emacs}
\alias{install_rfmt_emacs}
\title{Make the formatter available to Emacs, running under *nix, or 
Cygwin (on Windows).}
\usage{
install_rfmt_emacs(emacsrc = ".emacs", key = "C-x C-i",
  cygwin.path = NULL)
}
\arguments{
\item{emacsrc}{a string indicating Emacs initialization file in the user's 
home directory (by default, \code{.emacs}).}

\item{key}{the key combination used to invoke the formatter from Emacs
(by default, \dQuote{\code{Ctrl-x Ctrl-i}}).}

\item{cygwin.path}{if running Cygwin under Windows, the (Windows) path of the
Cygwin root directory (set to \dQuote{\code{C:/Cygwin}} if not provided).}
}
\description{
Make the formatter available to Emacs, running under *nix, or 
Cygwin (on Windows).
}

