% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfmt.R
\name{install_rfmt_shell}
\alias{install_rfmt_shell}
\title{Make the \dQuote{\code{rfmt}} command available to the *nix shell.}
\usage{
install_rfmt_shell(shellrc = ".bashrc", cygwin.path = NULL)
}
\arguments{
\item{shellrc}{a string indicating shell initialization file in the user's 
home directory (by default, \code{.bashrc}, the initialization file for the
\code{bash} shell).}

\item{cygwin.path}{if running \href{https://www.cygwin.com}{Cygwin} under 
Windows, the (Windows) path of the Cygwin root directory (set to 
\dQuote{\code{C:/Cygwin}} if not provided).}
}
\description{
Make the \dQuote{\code{rfmt}} command available to the *nix shell.
}

