% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfmt.R
\name{install_rfmt_vim}
\alias{install_rfmt_vim}
\title{Make the formatter available to \code{vim}, running under *nix, or 
Cygwin (on Windows).}
\usage{
install_rfmt_vim(vimrc = ".vimrc", key = "<C-I>", cygwin.path = NULL)
}
\arguments{
\item{vimrc}{a string indicating Emacs initialization file in the user's 
home directory (by default, \code{.vimrc}).}

\item{key}{the key combination used to invoke the formatter from \code{vim}
(by default, \dQuote{\code{Ctrl-I}}).}

\item{cygwin.path}{if running Cygwin under Windows, the (Windows) path of the
Cygwin root directory (set to \dQuote{\code{C:/Cygwin}} if not provided).}
}
\description{
Make the formatter available to \code{vim}, running under *nix, or 
Cygwin (on Windows).
}

