% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfmt.R
\name{rfmt}
\alias{rfmt}
\title{Format R code using the \code{rfmt} formatter}
\usage{
rfmt(source.file, text, ..., opts = list(...))
}
\arguments{
\item{source.file}{a character string giving the location of a file 
containing \R source code. Not supplied if parameter \code{text} is
given instead.}

\item{text}{a character string containing \R source code. Not supplied
if \code{source.file} is given instead.}

\item{opts}{options to \code{rfmt}, supplied as a named list.}

\item{\ldots}{options to \code{rfmt}. See \code{\link{options_rfmt}} for
options and their defaults.}
}
\value{
If formatting a file, nothing (\code{invisible(NULL)}). If formatting
a string, a \code{character} vector comprising the formatted code. If
formatting the contents of the clipboard, a similar \code{character} vector,
but rendered invisible; the formatted code is also output to the terminal
(using \code{cat}).
}
\description{
This function allows the \code{rfmt} \R source code formatting tool to be
invoked from \R. It applies \code{rfmt} to an \R source file, the contents
of the system clipboard, or to a character string.
}
\note{
Input is taken from the clipboard if neither \code{source.file} nor 
\code{text} is supplied.
}

