% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfmt.R
\name{rfmt_dir}
\alias{rfmt_dir}
\title{Format R code in a directory tree using the \code{rfmt} formatter}
\usage{
rfmt_dir(directory, pattern = "\\\\.[RrSs]$", all.files = FALSE,
  recursive = FALSE, opts = list())
}
\arguments{
\item{directory}{a character string giving the location of the directory
containing \R source code.}

\item{pattern}{a character string comprising a regular expression used to
distinguish \R source code files. By default, looks for files with an
extension "\code{r}" or "\code{s}" (lower- or upper case).}

\item{all.files}{logical, \code{TRUE} iff invisible files, as well as visible
files, are to be examined.}

\item{recursive}{logical, \code{TRUE} iff descendent directories of that 
specified by the parameter \code{directory} are to be examined.}

\item{opts}{options to \code{rfmt}, supplied as a named list.}
}
\description{
This function applies the the \code{rfmt} \R source code formatting
tool the \R source file(s) located in a specified directory and (optionally)
its descendents.
}

